feat(wire): int8 activation codec as default wire dtype + per-response tok/s badge
Add a self-contained int8 activation-wire codec (per-token abs-max symmetric
quantization with an optional keyframe-anchored delta pre-pass) behind a
header-self-describing dispatcher, so f32/f16/i8 negotiate per hop and the
orchestrator/host stay dtype-agnostic. Widen wireDtype to include 'i8' across
the planners, guards, and stage protocol; route hop-cost math through
legionActivationBytes (i8 = 1 byte/element). Default the wire to i8 (~4x fewer
bytes/token/hop than f32; overridable via ?wireDtype=f16 / ?wireDtype=f32).
i8 verified coherent on the 0.6b model: 4/5 prompts token-exact vs f32 under
greedy decode, the 5th a valid divergent completion (new f32-vs-i8 e2e A/B).

Surface a decode-throughput metric (first generated token -> last) from the
chat runtime and render it as a small "N tok/s" badge under each assistant
reply.

+++ packages/mesh-core/src/communalAssembly.ts
@@ -192,7 +192,7 @@ function syntheticRosterEntry(
           includeEmbeddings: claim.includeEmbeddings ?? claim.layerStart === 0,
           includeOutput: claim.includeOutput,
           ctxSize: 1,
-          wireDtype: 'f32',
+          wireDtype: 'i8',
           maxSessions,
           activeSessions: 0,
           epoch: 0,
